feat: organizando rotas para cada classe responsável
