+++ app/src/main/java/br/com/brunocarvalhs/howmuch/MainApp.kt
@@ -10,15 +10,15 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.EditProductRoute
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.FinalizePurchaseRoute
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.HistoryDetailRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.EditProductRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.FinalizePurchaseRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.HistoryDetailRoute
 import br.com.brunocarvalhs.howmuch.app.foundation.navigation.NavBarItem
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.ProductGraphRoute
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.SharedCartBottomSheetRoute
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.ShoppingCartGraphRoute
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.SubscriptionGraphRoute
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.TokenBottomSheetRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.ProductGraphRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.SharedCartBottomSheetRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.ShoppingCartGraphRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.SubscriptionGraphRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.TokenBottomSheetRoute
 import br.com.brunocarvalhs.howmuch.app.modules.base.BaseScreen
 import br.com.brunocarvalhs.howmuch.app.modules.editProduct.EditProductScreen
 import br.com.brunocarvalhs.howmuch.app.modules.finalizePurchase.FinalizePurchaseScreen
