--- app/src/main/java/br/com/brunocarvalhs/howmuch/app/foundation/navigation/HistoryDetailRoute.kt
@@ -1,6 +1,0 @@
-package br.com.brunocarvalhs.howmuch.app.foundation.navigation
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class HistoryDetailRoute(val cartId: String)
