+++ app/src/main/java/br/com/brunocarvalhs/howmuch/app/foundation/navigation/routes/FinalizePurchaseRoute.kt
@@ -1,0 +1,9 @@
+package br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class FinalizePurchaseRoute(
+    val cartId: String,
+    val price: Long,
+)
