+++ app/src/main/java/br/com/brunocarvalhs/howmuch/app/foundation/navigation/routes/SharedCartBottomSheetRoute.kt
@@ -1,0 +1,8 @@
+package br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SharedCartBottomSheetRoute(
+    val cartId: String,
+)
