+++ app/src/main/java/br/com/brunocarvalhs/howmuch/app/foundation/navigation/routes/TokenBottomSheetRoute.kt
@@ -1,0 +1,6 @@
+package br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TokenBottomSheetRoute(val token: String)
