+++ app/src/main/java/br/com/brunocarvalhs/howmuch/app/modules/finalizePurchase/FinalizePurchaseViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
 import br.com.brunocarvalhs.domain.usecases.cart.FinalizePurchaseUseCase
-import br.com.brunocarvalhs.howmuch.app.foundation.navigation.FinalizePurchaseRoute
+import br.com.brunocarvalhs.howmuch.app.foundation.navigation.routes.FinalizePurchaseRoute
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.MutableStateFlow
